+++ app/Http/Controllers/StatusController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Status;
+use Illuminate\Support\Facades\Auth;
 
 class StatusController extends Controller
 {
@@ -15,7 +16,8 @@ class StatusController extends Controller
     public function index()
     {
         $statuses = Status::all()->sortBy('id');
-        return view('status.index', compact('statuses'));
+        $isLoggedIn = Auth::check();
+        return view('status.index', compact('statuses', 'isLoggedIn'));
     }
 
     /**
@@ -26,6 +28,7 @@ class StatusController extends Controller
     public function create()
     {
         $status = new Status();
+        $this->authorize('create', $status);
         return view('status.create', compact('status'));
     }
 
@@ -42,7 +45,9 @@ class StatusController extends Controller
         ]);
         $status= new Status();
         $status->fill($data);
+        $this->authorize('store', $status);
         $status->save();
+        flash('Статус создан');
         return redirect()->route('statuses.index');
     }
 
@@ -66,6 +71,7 @@ class StatusController extends Controller
     public function edit($id)
     {
         $status = Status::findOrFail($id);
+        $this->authorize('edit', $status);
         return view('status.edit', compact('status'));
     }
 
@@ -83,7 +89,9 @@ class StatusController extends Controller
             'name' => 'required|unique:statuses,name,' . $status->id,
         ]);
         $status->fill($data);
+        $this->authorize('update', $status);
         $status->save();
+        flash('Статус обновлен');
         return redirect()->route('statuses.index');
     }
 
@@ -94,10 +102,14 @@ class StatusController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
-    {
+    { 
         $status = Status::find($id);
-        if ($status) {
+        $this->authorize('destroy', $status);
+        if ($status && $status->tasks->isEmpty()) {
           $status->delete();
+          flash('Статус удален');
+        } else {
+            flash('Статус не может быть удален')->error();
         }
         return redirect()->route('statuses.index');
     }
